+++ packages/xlsx-preview/src/xlcore_wasm.js
@@ -2727,6 +2727,31 @@ export class WorkbookHandle {
      * @param {any} update
      * @returns {any}
      */
+    updateImage(sheet, id, update) {
+        try {
+            const retptr = wasm.__wbindgen_add_to_stack_pointer(-16);
+            const ptr0 = passStringToWasm0(sheet, wasm.__wbindgen_export, wasm.__wbindgen_export2);
+            const len0 = WASM_VECTOR_LEN;
+            const ptr1 = passStringToWasm0(id, wasm.__wbindgen_export, wasm.__wbindgen_export2);
+            const len1 = WASM_VECTOR_LEN;
+            wasm.workbookhandle_updateImage(retptr, this.__wbg_ptr, ptr0, len0, ptr1, len1, addHeapObject(update));
+            var r0 = getDataViewMemory0().getInt32(retptr + 4 * 0, true);
+            var r1 = getDataViewMemory0().getInt32(retptr + 4 * 1, true);
+            var r2 = getDataViewMemory0().getInt32(retptr + 4 * 2, true);
+            if (r2) {
+                throw takeObject(r1);
+            }
+            return takeObject(r0);
+        } finally {
+            wasm.__wbindgen_add_to_stack_pointer(16);
+        }
+    }
+    /**
+     * @param {string} sheet
+     * @param {string} id
+     * @param {any} update
+     * @returns {any}
+     */
     updatePivot(sheet, id, update) {
         try {
             const retptr = wasm.__wbindgen_add_to_stack_pointer(-16);